fix(worker): 워커 프로세스그룹 취소 전파 보강
Constraint: 범위 밖 회귀 금지

🐙 Autopus <noreply@autopus.co>
