+++ pkg/worker/process_group_windows.go
@@ -1,0 +1,14 @@
+//go:build windows
+
+package worker
+
+import (
+	"context"
+	"os/exec"
+)
+
+func prepareCommandProcessGroup(cmd *exec.Cmd) {}
+
+func watchCommandCancellation(ctx context.Context, cmd *exec.Cmd, taskID string) func() {
+	return func() {}
+}
